Update instances of Card to ContentCard
