+++ src/docs/components/variants/contentCard/WithLink.tsx
@@ -1,21 +1,23 @@
 import Example from "../../../shared/components/Example";
-import Card from "../../../../library/components/Card";
+import ContentCard from "../../../../library/components/ContentCard";
 
-const DefaultCard = (): JSX.Element => {
+const WithLink = (): JSX.Element => {
   return (
     <Example>
-      <Card
+      <ContentCard
         texts={
           {
             cardTitle: "Card title",
             cardDescription: "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Duis nec vestibulum augue, viverra aliquet nunc. Cras eget felis sodales, vestibulum neque ac, rhoncus ipsum."
           }
         }
-        linkTitle={false}
-        headingLevel="h2"
+        headingLevel="h4"
+        linkHref="docs-demo-link"
+        actionLink={true}
+        actionLinkText="Call-to-action"
       />
     </Example>
   );
 }
 
-export default DefaultCard;
+export default WithLink;
