+++ src/docs/components/variants/contentCard/WithLinkedTitle.tsx
@@ -1,23 +1,22 @@
 import Example from "../../../shared/components/Example";
-import Card from "../../../../library/components/Card";
+import ContentCard from "../../../../library/components/ContentCard";
 
-const WithLink = (): JSX.Element => {
+const WithLinkedTitle = (): JSX.Element => {
   return (
     <Example>
-      <Card
+      <ContentCard
         texts={
           {
             cardTitle: "Card title",
             cardDescription: "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Duis nec vestibulum augue, viverra aliquet nunc. Cras eget felis sodales, vestibulum neque ac, rhoncus ipsum."
           }
         }
+        linkTitle={true}
         headingLevel="h4"
         linkHref="docs-demo-link"
-        actionLink={true}
-        actionLinkText="Call-to-action"
       />
     </Example>
   );
 }
 
-export default WithLink;
+export default WithLinkedTitle;
