+++ src/styles/components/_components.scss
@@ -5,7 +5,7 @@
 @import "accordion";
 @import "breadcrumbs";
 @import "button";
-@import "card";
+@import "content-card";
 @import "card-group";
 @import "dropdown";
 @import "link";
