+++ src/styles/components/_content-card.scss
